chore(union): post-rebase compatibility — k8s.NewClient signature, split oversized files
Rebasing onto main pulled in three changes that needed local adaptation:

* k8s.NewClient gained a kubeconfigDirs argument; updated completion.go
  callers to pass it through.
* commandbar_execute.go grew past the 800-line limit; extracted the
  kubectl argument helpers (commandAffectsNamespaces, injectKubectlDefaults,
  commandSupportsNamespaceFlags, kubectlGlobalFlagsWithValue,
  firstNonFlagToken) to commandbar_execute_kubectl_args.go.
* update_overlays_selectors.go grew past the 800-line limit; extracted
  the colorscheme overlay handlers to update_overlays_colorscheme.go.

No behavior change.
